# mobile/backend/docs/dating-api-spec.yaml
openapi: "3.0.3"
info:
  title: Dating API
  version: "1.0.0"
  description: |
    REST API для dating-приложения. Монорепо: backend (Go/chi) + frontend (React SPA).

    **Формат всех ответов:**
    ```json
    { "data": <payload | null>, "error": <"сообщение" | null> }
    ```

    **Rate limits:**
    - `/api/auth/*` — 5 rps, burst 10 (per IP)
    - `/api/*` (защищённые) — 100 rps, burst 200 (per IP)

    **Body limit:** 1 МБ на все запросы.
servers:
  - url: http://localhost:8080
    description: Локальная разработка
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access-токен, выданный при регистрации, логине или OAuth.
        TTL по умолчанию: 15 минут.
  schemas:
...
    Envelope:
      type: object
      required: [data, error]
      properties:
        data:
          description: Полезная нагрузка (null при ошибке)
        error:
          type: string
          nullable: true
          description: Сообщение об ошибке (null при успехе)

    # ── Пользователь ──────────────────────────────────────────────────────────
...
    ConversationResponse:
      type: object
      description: |
        Беседа в списке бесед. Поля name/avatar/lastMessage/timestamp/unread
        заполняются только в GET /api/conversations; в POST /api/conversations
        возвращается только id.
      required: [id, name, avatar, lastMessage, timestamp, unread]
      properties:
        id:
          type: string
          format: uuid
          example: "c1b2d3e4-0000-0000-0000-000000000001"
        name:
          type: string
          description: Имя собеседника (display_name партнёра)
          example: "Мария"
        avatar:
          type: string
          description: avatar_url партнёра
          example: "https://cdn.example.com/media/avatar.jpg"
        lastMessage:
          type: string
          description: Текст последнего сообщения (пустая строка если нет)
          example: "Привет! Как дела?"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601, UTC — момент последнего обновления беседы
          example: "2024-06-01T12:34:56Z"
        unread:
          type: boolean
          description: Есть ли непрочитанные сообщения
          example: true
    ConversationCreatedResponse:
      type: object
      description: Ответ на POST /api/conversations — только id беседы.
      required: [id]
      properties:
        id:
          type: string
          format: uuid
          example: "c1b2d3e4-0000-0000-0000-000000000001"
    MessageResponse:
      type: object
      description: Сообщение в беседе.
      required: [id, type, sender, time]
      properties:
        id:
          type: string
          format: uuid
          example: "m1b2d3e4-0000-0000-0000-000000000001"
        text:
          type: string
          nullable: true
          description: Текст сообщения (присутствует если type = "text")
          example: "Привет!"
        type:
          type: string
          enum: [text, image, video, voice]
          description: |
            Тип сообщения. Соответствует media_type_enum в БД;
            "text" если body != null и media_type is null.
          example: "text"
        sender:
          type: string
          enum: [me, other]
          description: '"me" если sender_id == текущий пользователь, иначе "other"'
          example: "me"
        time:
          type: string
          description: Время отправки в формате HH:MM (UTC)
          example: "14:05"
        mediaUrl:
          type: string
          nullable: true
          description: |
            URL медиафайла в CDN. Схема обязана быть https,
            хост должен совпадать с CDN_BASE_URL.
          example: "https://cdn.example.com/media/photo.jpg"
        duration:
          type: string
          nullable: true
          description: Длительность для voice/video — вычисляется на стороне клиента
          example: null

    # ── Storage ───────────────────────────────────────────────────────────────
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      description: Оболочка при ошибке.
      required: [data, error]
      properties:
        data:
          nullable: true
          example: null
        error:
          type: string
          example: "неверный email или пароль"
paths:
...
  /api/conversations:
    get:
      tags: [Conversations]
      summary: Список бесед текущего пользователя
      description: |
        Возвращает все беседы, в которых участвует авторизованный пользователь,
        с именем собеседника, аватаром, последним сообщением и флагом непрочитанных.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Список бесед
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ConversationResponse"
        "401":
          description: Отсутствует или недействителен JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Conversations]
      summary: Создать или получить прямую беседу с пользователем
      description: |
        Идемпотентная операция: если прямая беседа между текущим пользователем
        и `user_id` уже существует — возвращает её ID; иначе создаёт новую.
        В ответе возвращается только `id` беседы.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
                  format: uuid
                  description: UUID пользователя, с которым открывается беседа
                  example: "550e8400-e29b-41d4-a716-446655440001"
      responses:
        "200":
          description: ID беседы (новой или существующей)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ConversationCreatedResponse"
        "400":
          description: Неверный формат запроса или невалидный user_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Отсутствует или недействителен JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ════════════════════════════════════════════════════════════════════════════
  # MESSAGES
  # ════════════════════════════════════════════════════════════════════════════
  /api/conversations/{id}/messages:
    get:
      tags: [Messages]
      summary: История сообщений беседы
      description: |
        Возвращает сообщения беседы с пагинацией. Доступно только участникам беседы.
        Мягко удалённые сообщения (is_deleted = true) не возвращаются.
        По умолчанию: limit=20, page=1. Максимальный limit=100.
        Поле `sender` — "me" если sender_id == текущий пользователь, иначе "other".
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID беседы
          example: "c1b2d3e4-0000-0000-0000-000000000001"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Номер страницы (1-based)
          example: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Количество сообщений на странице
          example: 20
      responses:
        "200":
          description: Список сообщений
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Невалидный UUID беседы
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Отсутствует или недействителен JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Текущий пользователь не является участником беседы
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Messages]
      summary: Отправить сообщение в беседу
      description: |
        Создаёт новое сообщение в беседе. Доступно только участникам.
        Обязательно одно из полей: `text` или `mediaUrl`.

        **Ограничения для mediaUrl:**
        - Схема должна быть `https`
        - Хост должен совпадать с `CDN_BASE_URL` (SSRF-защита)

        **Ограничения для mediaType** (соответствует media_type_enum в БД):
        - Допустимые значения: `"image"`, `"video"`, `"voice"`
        - Если передан `mediaUrl`, рекомендуется передать `mediaType`

        CHECK-ограничение БД: `body IS NOT NULL OR media_url IS NOT NULL`.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID беседы
          example: "c1b2d3e4-0000-0000-0000-000000000001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Хотя бы одно из полей text или mediaUrl обязательно
              properties:
                text:
                  type: string
                  nullable: true
                  description: Текст сообщения
                  example: "Привет!"
                mediaUrl:
                  type: string
                  nullable: true
                  description: |
                    URL медиафайла (обычно fileUrl из /api/upload).
                    Схема https, хост = CDN_BASE_URL.
                  example: "https://cdn.example.com/media/photo.jpg"
                mediaType:
                  type: string
                  nullable: true
                  enum: [image, video, voice]
                  description: Тип медиафайла (обязателен при наличии mediaUrl)
                  example: "image"
      responses:
        "201":
          description: Сообщение создано
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MessageResponse"
        "400":
          description: |
            Неверный формат запроса, невалидный UUID беседы,
            отсутствуют text и mediaUrl, или mediaUrl нарушает ограничения CDN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Отсутствует или недействителен JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Текущий пользователь не является участником беседы
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/messages/{id}:
    delete:
      tags: [Messages]
      summary: Мягкое удаление сообщения
      description: |
        Устанавливает `is_deleted = true` для сообщения. Только автор сообщения
        может его удалить. Физически запись из БД не удаляется.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID сообщения
          example: "m1b2d3e4-0000-0000-0000-000000000001"
      responses:
        "200":
          description: Сообщение помечено как удалённое
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/OkResponse"
        "400":
          description: Невалидный UUID сообщения
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Отсутствует или недействителен JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Текущий пользователь не является автором сообщения
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ════════════════════════════════════════════════════════════════════════════
  # STORAGE
  # ════════════════════════════════════════════════════════════════════════════
...